+++ src/forze_mongo/kernel/platform/__init__.py
@@ -1,5 +1,6 @@
 from .client import MongoClient, MongoConfig, MongoTransactionOptions
+from .errors import mongo_handled
 
 # ----------------------- #
 
-__all__ = ["MongoClient", "MongoConfig", "MongoTransactionOptions"]
+__all__ = ["MongoClient", "MongoConfig", "MongoTransactionOptions", "mongo_handled"]